python-model/requirements.txt
# Base requirements for training and inference
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
Pillow>=9.0.0
numpy>=1.21.0
matplotlib>=3.5.0
tqdm>=4.64.0

# Web app
flask>=2.0.0
flask-cors>=3.0.0

# Computer vision
opencv-python>=4.5.0